moved fps.py to utils/fps.py (cleanup)
