+++ record_and_detect.py
@@ -1,14 +1,10 @@
 import cv2
 import datetime
-import json
-import os
-import time 
 from torch import Tensor as Tensor
 
 from copy import deepcopy
-from typing import Dict, List
 
-from fps import FPS
+from utils.fps import FPS
 from bohs_net_detector import BohsNetDetector
 from utils import save_to_json_file
